requirements.txt
playwright==1.31.1
pytest==7.4.2
pytest-allure-adaptor==2.9.0
python-dotenv==1.0.0
